Use 'ham' instead of 'not-spam'; make digrams
This is a Bad Commit -- two major changes.

First, I added code to make digram tokens -- tokens with pairs of words.

"I like ham" will generate the tokens ["I", "like", "ham", "^I",
"I^like", "like^ham", "ham^"]. The goal here is to have a little more
intelligence and context.

The other bit of code is that I changed "not-spam" as a type to
"ham". MongoDB, at least, was having problems with the "-" in the type
name.